+++ src/muir-utility.h
@@ -20,6 +20,6 @@ bool have_range(const MuirHD5 &file, boost::posix_time::time_period range);
 
 // Reads and parses the phasecode from an HDF5 file.
 // Return false if the the file doesn't contain a phasecode.
-bool read_phasecode(const MuirHD5 &file_in, std::vector<int> &phasecode);
+bool read_phasecode(const MuirHD5 &file_in, std::vector<float> &phasecode);
 
 #endif // MUIR_UTILITY
